Dirty hack to avoid issue with textarea
When the edit-disabled variable is set to an angular variable,
an issue appears with a textarea, where self.parent.addListeners.call(self);
is calling addListeners from the editableTextArea instance of
the editableDirectiveFactory. This causes a loop, freezing
the browser until it gives up when it runs out of stack.

This is a dirty workaround, because I can't figure out how to fix
the issue properly.